+++ client/backend/login.mjs
@@ -54,6 +54,8 @@ export async function login (call, db, username, password, { deviceId, deviceLab
         category: cat.value.category,
         isLive: cat.value.isLive,
         poster: cat.value.poster,
+        backdrop: cat.value.backdrop,
+        logo: cat.value.logo,
         feedKey: cat.value.feedKey,
         encryptionKey: b4a.toString(enc, 'hex')
       })